PruebaFlota
Cree Camion, Furgoneta y empecé a crear PruebaFlota

+++ Juan Esteban - Jacabo/src/Tipo1/Vehiculo.java
@@ -19,10 +19,9 @@ public abstract class Vehiculo {
         this.marca = marca;
     }
     
-    abstract float calcularAutonomia();
+    abstract double calcularAutonomia();
     String mostrarInformacionGeneral(){
         return "Vehiculo{" + "placa=" + placa + ", modelo=" + modelo + ", marca=" + marca + '}';
     }
 
-    
 }
